requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.0.3
grpcio==1.67.0
grpcio-tools==1.67.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpickle==3.3.0
MarkupSafe==3.0.2
pillow==11.0.0
protobuf==5.28.2
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.0.4
zipp==3.20.2
